BLD:Shippable Check if packages on conda; pip install if not
Having problems with nimfa being on the pip database but not on
conda. Annoyingly, conda doesn't revert to installing from pip if
a package can't be found on conda. This code will search the
database for the package and install it from conda if it is there,
and pip if it isn't.