BUG: deltaf For small data, prevent padlen exceeding length
If the data contains very few timepoints, our 30 taps mean the
automatic padlen used by filtfilt of `3 * ntaps` will exceed the
length of the data. We prevent this error by specifying the padlen
to the min of `3 * ntaps` and one less than the length of the data,
however further work needs to be done to give a better methodology
when the dataset is small.