MAINT: Use pool.join to clean up the pool
Usually, `pool.join()` is needed to ensure the parent process is
blocked until the pool is completed, however `pool.map` also
blocks, so we didn't need it for that.
https://docs.python.org/2/library/multiprocessing.html#multiprocessing.Process.join

However, `pool.join` ensures the pool is cleaned up properly and
exceptions are caught correctly.
https://stackoverflow.com/a/38271957/1960959

For us, we need to use `pool.join` so the coverage is measured
during each of the processes which are spawned, and they are
amalgamated together afterwards.
https://pytest-cov.readthedocs.io/en/latest/mp.html#ungraceful-pool-shutdown