REL: Specify maximum tifffile version for Py2.7/3.5
We need to specify this or pip will try to install a more
recent tifffile and imagecodecs version even though they have
dropped support for those versions of Python.

This is possibly a problem with tifffile, or its dependency
imagecodecs, not specifying which Python versions they support
correctly.

To do this, we use environment markers, as described here:
https://www.python.org/dev/peps/pep-0508/#environment-markers