CI: Don't change environment marker requirements in OLDEST_DEPS
When we specify
thing>=1.0.2; python_version>='3.6'
we need the `thing` to be set to ~=x.y.z (to try oldest supported
version), but not the `python_version` environment marker.

The solution here is to change all `>=` to `~=` in a global
replacement as before, but then find instances of `version~=` and
change them (back) to `version>=`
This is not the ideal solution, but sed does not support
conditional lookback statements, which would be needed to detect
a preceeding `;` in the line.