REL: Add past to requirements if python>=3
We removed this because it wasn't available on Python 2.7 and so
that was causing problems on Python 2.7. But we were actually
getting by because some of our dependencies happen require the
`past` package for their Python 3 implementations.

This uses the correct solution - requiring `past` conditional
on the Python version being >=3.