DOC: Add Binder configuration
Add a .binder directory, instructing binder to install our
requirements for the plots:

fissa[plotting]
sima

Note that this means that binder will be using the latest
stable release on pypi, not the bleeding-edge version of
fissa which is found in the master branch of the repository.

Because sima doesn't install on Python 3.7 and above, we also
specify the runtime to be Python 3.6.

specify binder python 3.6