API: datahandler is now a class
This involves quite a lot of changes.
This is classed as an API change as anyone using custom dataloaders in
our previous way will either have to stay with an older version of
FISSA, or migrate their datahandlers to this new method.
I will describe this in detail in the tutorial notebook that is to
follow in another commit.